phase3-chatbot/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.16
openai>=1.3.5
pydantic>=2.5.0
pydantic-settings>=2.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
slowapi>=0.1.9
python-dotenv>=1.0.0
alembic>=1.12.0
sqlalchemy>=2.0.0